Update to 1.1.1 - Networking
Since it was requested, the networking is now more functional and hopefully at least a tad easier.
