+++ Light-Show/src/net/exenco/lightshow/show/stage/fixtures/BlockUpdaterFixture.java
