+++ Light-Show/src/net/exenco/lightshow/show/stage/fixtures/FireworkFixture.java
