+++ Light-Show/src/net/exenco/lightshow/show/stage/fixtures/FogMachineFixture.java
