+++ Light-Show/src/net/exenco/lightshow/show/stage/fixtures/MovingHeadFixture.java
