+++ Light-Show/src/net/exenco/lightshow/show/stage/fixtures/ParticleFlareFixture.java
