+++ Light-Show/src/net/exenco/lightshow/show/stage/fixtures/ShowFixture.java
