+++ Light-Show/src/net/exenco/lightshow/util/ConfigHandler.java
