+++ Light-Show/src/net/exenco/lightshow/util/ProximitySensor.java
