+++ Light-Show/src/net/exenco/lightshow/util/VectorUtils.java
