+++ Light-Show/src/net/exenco/lightshow/util/api/LogoApi.java
