+++ Light-Show/src/net/exenco/lightshow/util/api/ParticleFlareApi.java
