+++ Light-Show/src/net/exenco/lightshow/util/api/ParticleLineApi.java
