+++ fixture-files/Exenco_Crystal.hed
@@ -1,0 +1,35 @@
+£Þ°“©׳††€‘ƒ‡™•ֻ€„‚ִֵָ ‘™­·÷´¯ת°¶»¦­ ע₪«מÿüשתי¿°±כ§«£¬³ֶֽ“ßװ”ּ׃
+כ–וֱֱׁÞÞ־ֲִ״ׂֱֳֻֿ…ֲּ׀וחûףÿפ÷
+ֿ´§¦£ס¿°ÜסטחמÝ«»«³
+׳×µ´³א­Ý»…™•™–§´„‡‡“ֳּׂ¨”„‡ֵÊְֶַֿ¢’¦­©½·רץ
+טחזארדגבאד‏‎üûזשרקציפףעסל—‰ˆ†…„ƒƒ€ָ־€›™˜—–••
+€וֹלךüערÿ»´§¦¥₪£¢¡¡£¾½¼»÷¹¸·×
+§װגל£׃־ֻּֽÊֱֹֻÚְֱֲֳִֵßÚֱ
+־»‹‡ָױִֹװ׃ׁׂ׀ןם¼נכךיטחזואÿ
+נ…¹£÷ןאûתשרקצצעןעסנ–
+›לÞÚֱ”€›™ָ„—–•”“’‘•³
+®­¬«¶
+©¨§¦¥₪£¢½
+ ¿¾½¼»÷¹₪·¶µ´³²±ֻּֽֿׂÊֶַָֹÙְֱֲֳִßÞÝְÛÚÙ״׳ײױװֿ
+ְֳּןמםלכךיטûזוהד‏באÿ‏בüûתשהקצץפןעסנ’
+‹‰ˆ‡™„ƒ‚‡™˜—
+•”“’‘«©±¬«×©´§¦¥₪¿¢¡ ¿¢
+½¼»÷¹¸³±©´³²±׃־ֻּֽײֶַָֹÙ
+ְֱֲֳִßÚÛְÛÚÙ״ֻײױװ׃־ׁ׀ןמס
+לכךיטחגדרדגבאד‏‎üûזשרקצי
+זבמ
+סנ‘™–››”—˜ƒ‚€ƒ›†™˜—–‰”“’‘¯®­¬·×©¨§¦¥₪£¾
+¡ ¿¾½¼÷÷¥
+×¥×§¦¯ £׃
+־׃ֻּÊַָֹÚֵÚְֱֲֳßÞֱÜֵÚÙ״׳ײױָ׃ּׁ׀ןמםלקךקטחזוהד‏ב‏ÿ‏‎üûתו
+ך £«¡±«¢אֿß׃ָׂײֽ›†…„ƒ
+‚€›†™„—•ˆ“‘¯®­¬«×µ
+¨»¦¹₪¿¢½ £¾¡¼§÷¥¸«¶©´¯
+²­ֽֿׂ׀
+ֻײ
+ֲֳִֵֶַָֹÝְßÞÝÜÛÚÙִ׳ײױװֿׁׂ׀ןעםלכךץטחזורדגבאד‏‎üûז
+שרקציפףעסל—‰ˆ‡†…„‡ּּ™ֹ™…˜‹
+„זײÚׁױֱף¼±
+¬·
+¸ג¨§¦¥₪£¢¡½¿¾½¼¦÷¹¸·«µ´³²¬ֿ־ֻּֽÊִֵֶַָֹ׀
+Ê
