--- src/plugin.yml
@@ -1,37 +1,0 @@
-name: Light-Show
-version: 1.1
-api-version: 1.18
-author: Exenco
-description: Display a Light-Show in Minecraft.
-main: net.exenco.lightshow.LightShow
-
-commands:
-  show:
-    description: Controls the plugin.
-    permission: lightshow.show
-    usage: /show
-
-permissions:
-  lightshow.*:
-    description: Gives access to all lightshow comamnds.
-    children:
-      lightshow.show: true
-      lightshow.reload: true
-      lightshow.start: true
-      lightshow.stop: true
-      lightshow.tos: true
-  lightshow.show:
-    description: Allows to control the lightshow.
-    default: true
-  lightshow.relaod:
-    description: Allows to reload the plugin.
-    default: op
-  lightshow.start:
-    description: Allows to start the ArtNet connection.
-    default: op
-  lightshow.stop:
-    description: Allows to stop the ArtNet connection.
-    default: op
-  lightshow.tos:
-    description: Allows to view Terms of Service.
-    default: true
